tests/wpt/mozilla/tests/bluetooth/requestDevice/canonicalizeFilter/empty-filters-member.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.glucose_heart_rate);
    return promise_rejects_js(t, TypeError, window.navigator.bluetooth.requestDevice({filters: []}),
                           'An empty |filters| member should result in a TypeError');
}, 'An empty |filters| member should result in a TypeError');
</script>
